taskProvider: Use the default hide property instead of the detail
We are using the detail content, if this was set to "hide", to hide
the task in the view. This is not the correct way to do it, because
tasks.json has a hide property. The issue is that the VS Code API
does not export this property to use it with vscode.Task. So, we
had to read the tasks.json file and get the hide property "manually".

Signed-off-by: Matheus Castello <matheus@castello.eng.br>